+++ android-sdk-ui/src/com/appboy/ui/inappmessage/views/AppboyInAppMessageHtmlFullView.java
@@ -8,12 +8,20 @@ import com.appboy.ui.R;
 
 public class AppboyInAppMessageHtmlFullView extends AppboyInAppMessageHtmlBaseView {
 
+  private WebView mMessageWebView;
+
   public AppboyInAppMessageHtmlFullView(Context context, AttributeSet attrs) {
     super(context, attrs);
   }
 
   @Override
   public WebView getMessageWebView() {
-    return (WebView) findViewById(R.id.com_appboy_inappmessage_html_full_webview);
+    if (mMessageWebView == null) {
+      mMessageWebView = (WebView) findViewById(R.id.com_appboy_inappmessage_html_full_webview);
+      if (mMessageWebView != null) {
+        mMessageWebView.getSettings().setJavaScriptEnabled(true);
+      }
+    }
+    return mMessageWebView;
   }
 }