+++ droidboy/src/com/appboy/sample/InAppMessageTesterActivity.java
@@ -36,6 +36,11 @@ import java.util.HashMap;
 import java.util.Map;
 
 public class InAppMessageTesterActivity extends AppboyFragmentActivity implements AdapterView.OnItemSelectedListener {
+
+  private enum HtmlMessageJsType {
+    NO_JS, INLINE_JS, EXTERNAL_JS
+  }
+
   private static final String CUSTOM_INAPPMESSAGE_VIEW_KEY = "inapmessages_custom_inappmessage_view";
   private static final String CUSTOM_INAPPMESSAGE_MANAGER_LISTENER_KEY = "inappmessages_custom_inappmessage_manager_listener";
   private static final String CUSTOM_APPBOY_NAVIGATOR_KEY = "inappmessages_custom_appboy_navigator";
@@ -58,7 +63,8 @@ public class InAppMessageTesterActivity extends AppboyFragmentActivity implement
     "We don't recommend making in-app messages longer than 140 characters due to variations in screens.  This is an in-app message & this message is exactly six hundred and forty chars!  " +
     "We don't recommend making in-app messages longer than 140 characters due to variations in screens.  This is a waaaay too long message#";
 
-  private static final String HTML_CONTENT_REMOTE_URL = "https://s3.amazonaws.com/appboy-development-bucket/droidboy_html_tester_zipped_assets.zip";
+  private static final String HTML_ASSETS_NO_JS_REMOTE_URL = "https://s3.amazonaws.com/appboy-development-bucket/droidboy_html_tester_zipped_assets.zip";
+  private static final String HTML_ASSETS_WITH_EXTERNAL_JS_REMOTE_URL = "https://www.dropbox.com/s/ikn7ge2tfrfx2cn/Archive.zip?dl=1";
 
   // color reference: http://www.google.com/design/spec/style/color.html
   private static final int APPBOY_RED = 0xFFf33e3e;
@@ -111,6 +117,8 @@ public class InAppMessageTesterActivity extends AppboyFragmentActivity implement
   private String mImage;
   private String mButtons;
   private String mHtmlBodyFromAssets;
+  private String mHtmlBodyFromAssetsInlineJS;
+  private String mHtmlBodyFromAssetsExternalJS;
 
   @Override
   public void onCreate(Bundle savedInstanceState) {
@@ -180,8 +188,12 @@ public class InAppMessageTesterActivity extends AppboyFragmentActivity implement
             addInAppMessage(new InAppMessageModal());
           } else if ("full".equals(mMessageType)) {
             addInAppMessage(new InAppMessageFull());
-          } else if ("html_full".equals(mMessageType)) {
-            addInAppMessage(new InAppMessageHtmlFull());
+          } else if ("html_full_no_js".equals(mMessageType)) {
+            addInAppMessage(new InAppMessageHtmlFull(), HtmlMessageJsType.NO_JS);
+          } else if ("html_full_inline_js".equals(mMessageType)) {
+            addInAppMessage(new InAppMessageHtmlFull(), HtmlMessageJsType.INLINE_JS);
+          } else if ("html_full_external_js".equals(mMessageType)) {
+            addInAppMessage(new InAppMessageHtmlFull(), HtmlMessageJsType.EXTERNAL_JS);
           } else {
             addInAppMessage(new InAppMessageSlideup());
           }
@@ -242,7 +254,9 @@ public class InAppMessageTesterActivity extends AppboyFragmentActivity implement
       }
     });
 
-    mHtmlBodyFromAssets = readHtmlBodyFromAssets();
+    mHtmlBodyFromAssets = readHtmlBodyFromAssets(HtmlMessageJsType.NO_JS);
+    mHtmlBodyFromAssetsInlineJS = readHtmlBodyFromAssets(HtmlMessageJsType.INLINE_JS);
+    mHtmlBodyFromAssetsExternalJS = readHtmlBodyFromAssets(HtmlMessageJsType.EXTERNAL_JS);
   }
 
   private void addInAppMessageImmersive(IInAppMessageImmersive inAppMessage) {
@@ -280,18 +294,35 @@ public class InAppMessageTesterActivity extends AppboyFragmentActivity implement
     inAppMessage.setIcon("\uf091");
   }
 
-  private void addInAppMessageHtmlFull(IInAppMessageHtml inAppMessage) {
-    inAppMessage.setMessage(mHtmlBodyFromAssets);
-    inAppMessage.setAssetsZipRemoteUrl(HTML_CONTENT_REMOTE_URL);
+  private void addInAppMessageHtmlFull(IInAppMessageHtml inAppMessage, HtmlMessageJsType jsType) {
+    switch (jsType) {
+      case NO_JS:
+        inAppMessage.setMessage(mHtmlBodyFromAssets);
+        inAppMessage.setAssetsZipRemoteUrl(HTML_ASSETS_NO_JS_REMOTE_URL);
+        break;
+      case INLINE_JS:
+        inAppMessage.setMessage(mHtmlBodyFromAssetsInlineJS);
+        inAppMessage.setAssetsZipRemoteUrl(HTML_ASSETS_WITH_EXTERNAL_JS_REMOTE_URL);
+        break;
+      case EXTERNAL_JS:
+        inAppMessage.setMessage(mHtmlBodyFromAssetsExternalJS);
+        inAppMessage.setAssetsZipRemoteUrl(HTML_ASSETS_WITH_EXTERNAL_JS_REMOTE_URL);
+        break;
+      default:
+        break;
+    }
   }
 
   private void addInAppMessage(IInAppMessage inAppMessage) {
+    addInAppMessage(inAppMessage, null);
+  }
+  private void addInAppMessage(IInAppMessage inAppMessage, HtmlMessageJsType jsType) {
     if (inAppMessage instanceof IInAppMessageImmersive) {
       addInAppMessageImmersive((IInAppMessageImmersive) inAppMessage);
     } else if (inAppMessage instanceof InAppMessageSlideup) {
       addInAppMessageSlideup((InAppMessageSlideup) inAppMessage);
     } else if (inAppMessage instanceof InAppMessageHtmlFull) {
-      addInAppMessageHtmlFull((InAppMessageHtmlFull) inAppMessage);
+      addInAppMessageHtmlFull((InAppMessageHtmlFull) inAppMessage, jsType);
     } else if (inAppMessage instanceof IInAppMessage) {
       addInAppMessageCustom(inAppMessage);
     }
@@ -560,12 +591,27 @@ public class InAppMessageTesterActivity extends AppboyFragmentActivity implement
   /**
    * @return the html body string from the assets folder or null if the read fails.
    */
-  private String readHtmlBodyFromAssets() {
+  private String readHtmlBodyFromAssets(HtmlMessageJsType jsType) {
+    return readHtmlBodyFromAssetsWithFileName(jsType);
+  }
+
+  private String readHtmlBodyFromAssetsWithFileName(HtmlMessageJsType jsType) {
     String htmlBody = null;
+    String filename = "html_inapp_message_body_no_js.html";
+    switch (jsType) {
+      case INLINE_JS:
+        filename = "html_inapp_message_body_inline_js.html";
+        break;
+      case EXTERNAL_JS:
+        filename = "html_inapp_message_body_external_js.html";
+        break;
+      default:
+        break;
+    }
 
     // Get the text of the html from the assets folder
     try {
-      BufferedReader reader = new BufferedReader(new InputStreamReader(getAssets().open("html_inapp_message_body.txt"), "UTF-8"));
+      BufferedReader reader = new BufferedReader(new InputStreamReader(getAssets().open(filename), "UTF-8"));
       String line;
       StringBuilder stringBuilder = new StringBuilder();
       while ((line = reader.readLine()) != null) {