+++ droidboy/src/com/appboy/sample/UserProfileDialog.java
@@ -28,14 +28,12 @@ public class UserProfileDialog extends DialogPreference {
   private static final String FIRST_NAME_PREFERENCE_KEY = "user.firstname";
   private static final String LAST_NAME_PREFERENCE_KEY = "user.lastname";
   private static final String EMAIL_PREFERENCE_KEY = "user.email";
-  private static final String BIO_PREFERENCE_KEY = "user.bio";
   private static final String GENDER_PREFERENCE_KEY = "user.gender_resource_id";
   private static final String AVATAR_PREFERENCE_KEY = "user.avatar_image_url";
 
   private static final String SAMPLE_FIRST_NAME = "Jane";
   private static final String SAMPLE_LAST_NAME = "Doe";
   private static final String SAMPLE_EMAIL = "jane@appboy.com";
-  private static final String SAMPLE_BIO = "I'm a Developer";
   private static final int SAMPLE_GENDER = R.id.female;
   private static final String SAMPLE_AVATAR_URL = "https://s3.amazonaws.com/appboy-dashboard-uploads/news/default-news-image.png";
 
@@ -43,7 +41,6 @@ public class UserProfileDialog extends DialogPreference {
   private EditText mFirstName;
   private EditText mLastName;
   private EditText mEmail;
-  private EditText mBio;
   private RadioGroup mGender;
   private EditText mAvatarImageUrl;
   private CheckBox mRequestFlush;
@@ -66,7 +63,6 @@ public class UserProfileDialog extends DialogPreference {
     mFirstName = (EditText) view.findViewById(R.id.first_name);
     mLastName = (EditText) view.findViewById(R.id.last_name);
     mEmail = (EditText) view.findViewById(R.id.email);
-    mBio = (EditText) view.findViewById(R.id.bio);
     mGender = (RadioGroup) view.findViewById(R.id.gender);
     mAvatarImageUrl = (EditText) view.findViewById(R.id.avatar_image_url);
     mRequestFlush = (CheckBox) view.findViewById(R.id.user_dialog_flush_checkbox);
@@ -81,14 +77,12 @@ public class UserProfileDialog extends DialogPreference {
     mFirstName.setText(sharedPreferences.getString(FIRST_NAME_PREFERENCE_KEY, null));
     mLastName.setText(sharedPreferences.getString(LAST_NAME_PREFERENCE_KEY, null));
     mEmail.setText(sharedPreferences.getString(EMAIL_PREFERENCE_KEY, null));
-    mBio.setText(sharedPreferences.getString(BIO_PREFERENCE_KEY, null));
     mGender.check(parseGenderFromSharedPreferences());
     mAvatarImageUrl.setText(sharedPreferences.getString(AVATAR_PREFERENCE_KEY, null));
     mRequestFlush.setChecked(false);
     ButtonUtils.setUpPopulateButton(view, R.id.first_name_button, mFirstName, getSharedPreferences().getString(FIRST_NAME_PREFERENCE_KEY, SAMPLE_FIRST_NAME));
     ButtonUtils.setUpPopulateButton(view, R.id.last_name_button, mLastName, getSharedPreferences().getString(LAST_NAME_PREFERENCE_KEY, SAMPLE_LAST_NAME));
     ButtonUtils.setUpPopulateButton(view, R.id.email_button, mEmail, getSharedPreferences().getString(EMAIL_PREFERENCE_KEY, SAMPLE_EMAIL));
-    ButtonUtils.setUpPopulateButton(view, R.id.bio_button, mBio, getSharedPreferences().getString(BIO_PREFERENCE_KEY, SAMPLE_BIO));
     ButtonUtils.setUpPopulateButton(view, R.id.avatar_image_url_button, mAvatarImageUrl, getSharedPreferences().getString(AVATAR_PREFERENCE_KEY, SAMPLE_AVATAR_URL));
 
     final Button populateButton = (Button) view.findViewById(R.id.user_dialog_button_populate);
@@ -102,9 +96,6 @@ public class UserProfileDialog extends DialogPreference {
         }
         if (mEmail.getText().length() == 0) {
           mEmail.setText(getSharedPreferences().getString(EMAIL_PREFERENCE_KEY, SAMPLE_EMAIL));
-        }
-        if (mBio.getText().length() == 0) {
-          mBio.setText(getSharedPreferences().getString(BIO_PREFERENCE_KEY, SAMPLE_BIO));
         }
         if (mGender.getCheckedRadioButtonId() == R.id.unspecified) {
           mGender.check(SAMPLE_GENDER);
@@ -120,7 +111,6 @@ public class UserProfileDialog extends DialogPreference {
         mFirstName.getText().clear();
         mLastName.getText().clear();
         mEmail.getText().clear();
-        mBio.getText().clear();
         mGender.check(R.id.unspecified);
         mAvatarImageUrl.getText().clear();
       }
@@ -133,7 +123,6 @@ public class UserProfileDialog extends DialogPreference {
       String firstName = mFirstName.getText().toString();
       String lastName = mLastName.getText().toString();
       String email = mEmail.getText().toString();
-      String bio = mBio.getText().toString();
       int genderResourceId = mGender.getCheckedRadioButtonId();
       View genderRadioButton = mGender.findViewById(genderResourceId);
       int genderId = mGender.indexOfChild(genderRadioButton);
@@ -149,7 +138,6 @@ public class UserProfileDialog extends DialogPreference {
       editor.putString(FIRST_NAME_PREFERENCE_KEY, firstName);
       editor.putString(LAST_NAME_PREFERENCE_KEY, lastName);
       editor.putString(EMAIL_PREFERENCE_KEY, email);
-      editor.putString(BIO_PREFERENCE_KEY, bio);
       editor.putInt(GENDER_PREFERENCE_KEY, genderId);
       editor.putString(AVATAR_PREFERENCE_KEY, avatarImageUrl);
       editor.apply();
@@ -158,7 +146,6 @@ public class UserProfileDialog extends DialogPreference {
       appboyUser.setFirstName(firstName);
       appboyUser.setLastName(lastName);
       appboyUser.setEmail(email);
-      appboyUser.setBio(bio);
       switch (genderId) {
         case GENDER_UNSPECIFIED_INDEX:
           appboyUser.setGender(null);